requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
albumentations>=1.3.1
torchmetrics>=1.0.0
numpy<2
Pillow>=9.5.0
opencv-python
boxmot
tqdm
supervision
gradio
imageio
imageio-ffmpeg
